refactor(retrieval): move the read path out of the service
internal/service/memory.go was 1,930 lines holding protocol translation,
three-retriever orchestration, extraction, fold, entity matching and
profile assembly in one type. The read path is the part of that hardest to
reason about and easiest to break, and none of it could be exercised
without a gRPC request around it.

internal/retrieval now owns what comes back and in what order: query
parsing and keyword extraction, the candidate pool sizing that decides what
ranking is allowed to consider, the three retrievers, the fallback for a
query with nothing to search for, the merge, and the rank. The service
keeps what serving a request needs -- protobuf translation, context edges,
access counts, metrics -- and the two now fail differently: retrieval
returns ordinary errors and the service is what turns them into gRPC
status codes.

The Repo interface names six operations out of the repository's several
dozen. Per ADR 0001 that is a test seam, not a portability claim, and
nothing here is shaped to keep another graph store swappable.

Behaviour is unchanged, and the golden suite is how that is known rather
than asserted: recall@10 0.909 and MRR 0.828 before, identical after, with
the per-group numbers matching to three decimals.

Two stranded comment fragments went with it -- leftovers from an earlier
edit, sitting above a const block and describing functions defined
elsewhere.

Closes #64
